+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### 1.0.0: October 22, 2021
 * Updating version in package.json, and changed package name to landuseplanning-admin
 * Survey choices restriction (LUP-240)
+* Recent activity bug not showing on homepage, after update (LUP-241)